requirements.txt
gpytranslate==2.0.0
openai==1.57.3
python-dotenv==1.0.1
aiohttp==3.10
aiogram==3.15.0
psycopg-binary==3.2.3
psycopg==3.2.3
psycopg-pool==3.2.6
aiocryptopay==0.4.7
tiktoken==0.8.0
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.9
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
telethon==1.34.0
# asyncio - это стандартная библиотека Python, не нужно устанавливать отдельно